Keylogger007/requirements.txt
pynput>=1.7.6 
Pillow>=8.0.0
